Move blocking parts to blocking module
