+++ atat/src/blocking/client.rs
@@ -4,8 +4,9 @@ use fugit::ExtU32;
 
 use crate::error::{Error, Response};
 use crate::helpers::LossyStr;
-use crate::traits::{AtatClient, AtatCmd, AtatUrc};
-use crate::Config;
+use crate::{Config, AtatUrc};
+
+use super::{AtatCmd, AtatClient};
 
 #[derive(Debug, PartialEq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
